+++ src/IO/Win32/FileUtils_win32.cpp
@@ -160,7 +160,7 @@ void FileUtils::Delete(const wchar_t* filePath)
 //-----------------------------------------------------------------------------
 //
 //-----------------------------------------------------------------------------
-size_t FileUtils::GetFileSize(const TCHAR* filePath)
+uint64_t FileUtils::GetFileSize(const TCHAR* filePath)
 {
 	LN_THROW( filePath != NULL, ArgumentException );
 
@@ -168,7 +168,7 @@ size_t FileUtils::GetFileSize(const TCHAR* filePath)
 	errno_t r = _tfopen_s( &fp, filePath, _T("r") );
 	LN_THROW( r == 0, FileNotFoundException );
 
-	size_t size = 0;
+	uint64_t size = 0;
 	try
 	{
 		size = GetFileSize(fp);
@@ -185,7 +185,7 @@ size_t FileUtils::GetFileSize(const TCHAR* filePath)
 //-----------------------------------------------------------------------------
 //
 //-----------------------------------------------------------------------------
-size_t FileUtils::GetFileSize( FILE* stream )
+uint64_t FileUtils::GetFileSize(FILE* stream)
 {
 	struct _stat stbuf;
 	int handle = _fileno( stream );
